Refactor chip offer endpoint to call utils function for offer-writing logic
